Allow overrides plus public/private commands.
This commit refactors the subcommand logic to allow for multiple
definitions of commands (last in wins) and also does not error out when
a subcommand file is not found.

The Import field for the Cmd struct can now handle multiple paths as in

Import: |
  path/1
  path/2
  path/3

If a command in path/2 is already defined in a command in path/1 then
the latter will be used.

Also if path/3 does not exist the user can still work with commands
defined in path/1 and path/2.  By not throwing an error in this case, we
allow the ability to publicly share a config for ahoy commands that
points to private inaccessible areas without erroring out.

This commit also include some new examples that illustrate this
functionality.