Allow include command files.
As an ahoy user I wish to be able to organize long commands into their
own file.

Also as a user I wish to able to override some default command without
editing an upstream config or command file.

This commit allows users to specify a list of include files that are
consumed by ahoy.

Also, this commit lays the groundwork for being able to override
upstream command definitions.