feature: Add command aliases as requested in #110
This feature adds an optional `"aliases"` array parameter to each
command in the `.ahoy.yml` file.

This must be an array of one or more strings.

If more than one command has the same alias, the "last in wins"
rule will be used, meaning the last command found by the Yaml
parser (usually the one at the bottom of the file) will be run.

Signed-off-by: Drew Robinson <drew.robinson@gmail.com>